refactor: Moved the repositoryTest class file to the repository package.
