+++ back/deukgeun/src/test/java/com/dg/deukgeun/repository/WorkoutSessionRepositoryTest.java
@@ -1,10 +1,7 @@
-package com.dg.deukgeun;
+package com.dg.deukgeun.repository;
 
-import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-
-import com.dg.deukgeun.repository.WorkoutSessionRepository;
 
 import lombok.extern.log4j.Log4j2;
